feat(tti): rename
