+++ packages/falcon-ecommerce-uikit/src/EnsureTTI/index.tsx
@@ -1,0 +1,1 @@
+export * from './EnsureTTI';
