--- packages/falcon-ecommerce-uikit/src/Idle/index.tsx
@@ -1,1 +1,0 @@
-export * from './Idle';
