+++ packages/falcon-ecommerce-uikit/src/index.ts
@@ -27,4 +27,4 @@ export * from './Forms';
 export * from './AccountRecovery';
 export * from './Errors';
 export * from './SortOrdersProvider';
-export * from './Idle';
+export * from './EnsureTTI';
